:sparkles: Add the elementIsInView function to the lib
The function will take an element and use its getBoundingClientRect values to determine whether or not is it in view of the user. The dev can pass an optional second parameter to tell the function to limit returning true only to when the full element is in view. By default the function will return true if the element is partially visible.

Added accompanying Jest test which is passing